requirements.txt
dash==2.11.1
plotly==5.15.0
pandas==2.0.2
numpy==1.24.2
pymongo==4.4.1
python-dotenv==1.0.0